Remove .php and .json files, and consolidate to one directory
