+++ packages/styled-components/components/app/component.json
