+++ packages/styled-components/components/app/index.js
