+++ packages/styled-components/components/byline/themes/default.js
