--- packages/styled-components/components/common/html/component.json
@@ -1,31 +1,0 @@
-{
-	"name": "html",
-	"category": "Common",
-	"description": "Use `dangerouslySetInnerHTML` to render HTML.",
-	"theme_options": [
-		"default",
-		"unstyled"
-	],
-	"config": {
-		"content": {
-			"default": "",
-			"required": true,
-			"type": "string"
-		},
-		"oembed": {
-			"default": false,
-			"required": true,
-			"type": "bool"
-		},
-		"style": {
-			"default": [],
-			"required": true,
-			"type": "array"
-		},
-		"tag": {
-			"default": "div",
-			"required": true,
-			"type": "string"
-		}
-	}
-}
