--- packages/styled-components/components/common/link/component.json
@@ -1,27 +1,0 @@
-{
-	"name": "irving/link",
-	"category": "Common",
-	"description": "Custom anchor.",
-	"config": {
-		"href": {
-			"default": "",
-			"required": true,
-			"type": "string"
-		},
-		"rel": {
-			"default": "",
-			"required": true,
-			"type": "string"
-		},
-		"style": {
-			"default": [],
-			"required": true,
-			"type": "array"
-		},
-		"target": {
-			"default": "",
-			"required": true,
-			"type": "string"
-		}
-	}
-}
