+++ packages/styled-components/components/featuredMedia/component.json
