+++ packages/styled-components/components/featuredMedia/index.php
