+++ packages/styled-components/components/fragment/themes/default.js
