+++ packages/styled-components/components/html/themes/default.js
