--- packages/styled-components/components/layouts/fragment/component.json
@@ -1,11 +1,0 @@
-{
-	"name": "irving/fragment",
-	"category": "Common",
-	"description": "Render children in a React fragment or any other HTML tag.",
-	"config": {
-		"tag": {
-			"type": "string",
-			"required": false
-		}
-	}
-}
