+++ packages/styled-components/components/logo/component.json
