+++ packages/styled-components/components/logo/index.js
