+++ packages/styled-components/components/logo/index.php
