+++ packages/styled-components/components/logo/themes/default.js
