+++ packages/styled-components/components/menu/component.json
