+++ packages/styled-components/components/menu/themes/defaultVertical.js
