--- packages/styled-components/components/wordpress/page/archive-title/component.json
@@ -1,6 +1,0 @@
-{
-	"name": "irving/archive-title",
-	"category": "Archive",
-	"description": "Get the archive title",
-	"config": {}
-}
