--- packages/styled-components/components/wordpress/post/excerpt/component.json
@@ -1,20 +1,0 @@
-{
-	"name": "irving/post-excerpt",
-	"category": "Post",
-	"description": "Get the post excerpt",
-	"config": {
-		"content": {
-			"default": "",
-			"required": true,
-			"type": "string"
-		},
-		"post_id": {
-			"default": 0,
-			"required": true,
-			"type": "integer"
-		}
-	},
-	"use_context": {
-		"irving/post_id": "post_id"
-	}
-}
