--- packages/styled-components/components/wordpress/post/list/component.json
@@ -1,13 +1,0 @@
-{
-	"name": "irving/post-list",
-	"description": "Iterate through a WP_Query's posts.",
-	"config": {
-		"templates": {
-			"type": "array"
-		},
-		"query_args": {
-			"type": "array",
-			"required": false
-		}
-	}
-}
