--- packages/styled-components/components/wordpress/post/post/component.json
@@ -1,15 +1,0 @@
-{
-	"name": "irving/post",
-	"category": "Data",
-	"description": "Post context provider.",
-	"config": {
-		"post_id": {
-			"default": 0,
-			"required": true,
-			"type": "integer"
-		}
-	},
-	"provides_context": {
-		"irving/post_id": "post_id"
-	}
-}
