--- packages/styled-components/components/wordpress/site/bloginfo/component.json
@@ -1,17 +1,0 @@
-{
-	"name": "irving/bloginfo",
-	"category": "Site",
-	"description": "Easily call bloginfo()",
-	"config": {
-		"content": {
-			"default": "",
-			"required": true,
-			"type": "string"
-		},
-		"show": {
-			"default": "name",
-			"required": true,
-			"type": "string"
-		}
-	}
-}
